CLI adding series and cards from json

+++ src/card.rs
@@ -3,7 +3,7 @@ pub struct Card {
     pub name: String,
     pub series_id: i32,
     pub number: String,
-    pub collection_number: i16,
+    pub collection_number: i32,
     pub in_collection: bool,
     pub rarity_id: i32,
 }
